更新数据与主干保持一致

+++ iPad Viewer/DataStorage/LoggerDataStorage.m
@@ -5,38 +5,30 @@
  * Copyright (c) 2012-2013 Sung-Taek, Kim <stkim1@colorfulglue.com> All Rights
  * Reserved.
  *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
+ * Redistribution and use in source and binary forms, with or without modification,
+ * are permitted provided that the following conditions are met:
  *
- * 1. Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * 3. Any redistribution is done solely for personal benefit and not for any
- *    commercial purpose or for monetary gain
- *
- * 4. No binary form of source code is submitted to App Store℠ of Apple Inc.
- *
- * 5. Neither the name of the Sung-Taek, Kim nor the names of its contributors
- *    may be used to endorse or promote products derived from  this software
- *    without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR  IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL COPYRIGHT HOLDER AND AND CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
+ * Redistributions of  source code  must retain  the above  copyright notice,
+ * this list of  conditions and the following  disclaimer. Redistributions in
+ * binary  form must  reproduce  the  above copyright  notice,  this list  of
+ * conditions and the following disclaimer  in the documentation and/or other
+ * materials  provided with  the distribution.  Neither the  name of Sung-Tae
+ * k Kim nor the names of its contributors may be used to endorse or promote
+ * products  derived  from  this  software  without  specific  prior  written
+ * permission.  THIS  SOFTWARE  IS  PROVIDED BY  THE  COPYRIGHT  HOLDERS  AND
+ * CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT
+ * NOT LIMITED TO, THE IMPLIED  WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A  PARTICULAR PURPOSE  ARE DISCLAIMED.  IN  NO EVENT  SHALL THE  COPYRIGHT
+ * HOLDER OR  CONTRIBUTORS BE  LIABLE FOR  ANY DIRECT,  INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY,  OR CONSEQUENTIAL DAMAGES (INCLUDING,  BUT NOT LIMITED
+ * TO, PROCUREMENT  OF SUBSTITUTE GOODS  OR SERVICES;  LOSS OF USE,  DATA, OR
+ * PROFITS; OR  BUSINESS INTERRUPTION)  HOWEVER CAUSED AND  ON ANY  THEORY OF
+ * LIABILITY,  WHETHER  IN CONTRACT,  STRICT  LIABILITY,  OR TORT  (INCLUDING
+ * NEGLIGENCE  OR OTHERWISE)  ARISING  IN ANY  WAY  OUT OF  THE  USE OF  THIS
+ * SOFTWARE,   EVEN  IF   ADVISED  OF   THE  POSSIBILITY   OF  SUCH   DAMAGE.
  *
  */
+
 
 
 #import "LoggerDataStorage.h"
@@ -62,7 +54,7 @@
 @property (nonatomic, readonly) NSMutableArray			*readOperationSlot;
 
 @property (nonatomic, retain)	NSString				*basepath;
-@property (nonatomic, readonly) int						cpuCount;
+@property (nonatomic, readonly) NSUInteger				cpuCount;
 
 
 -(void)_purgeDataEntryCache;
@@ -103,14 +95,14 @@ unsigned int _delete_dependency_count(NSArray*, LoggerDataDelete*);
 	dispatch_queue_t		_highPriorityOperationQueue;
 
 	// according to this count, we will resume or suspend the queue above
-	unsigned int			_high_queue_suspension_count;
+	NSUInteger			_high_queue_suspension_count;
 
 	// this queue will concurrently dispatch actual instructions
 	dispatch_queue_t		_operationDispatcherQueue;
 
 	// this is the cache pool that contains binary data
 	NSMutableDictionary		*_dataEntryCache;
-	unsigned int			_dataEntryCacheSize;
+	NSUInteger			_dataEntryCacheSize;
 	
 	// data operation queue
 	NSMutableArray			*_operationPool;
@@ -125,7 +117,7 @@ unsigned int _delete_dependency_count(NSArray*, LoggerDataDelete*);
 	// a path where directory operation must based on
 	NSString				*_basepath;
 	
-	int						_cpuCount;
+	NSUInteger				_cpuCount;
 }
 @synthesize lowPriorityOperationQueue			= _lowPriorityOperationQueue;
 @synthesize highPriorityOperationQueue			= _highPriorityOperationQueue;
@@ -171,7 +163,7 @@ SYNTHESIZE_SINGLETON_FOR_CLASS_WITH_ACCESSOR(LoggerDataStorage,sharedDataStorage
 		dispatch_set_target_queue(_lowPriorityOperationQueue,_highPriorityOperationQueue);
 		
 		// cpu core counts
-		int cpus = [[NSProcessInfo processInfo] processorCount];
+		NSUInteger cpus = [[NSProcessInfo processInfo] processorCount];
 		
 		dispatch_sync(_highPriorityOperationQueue, ^{
